update folder name, useState var, and convertGql..
