+++ MockQuery_Tools/insertEntriesIntoObject.js
